SKALE-1755 refactored BLS code. moved into new JS module file
